requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
fpyutils==1.1.0
github3.py==1.3.0
idna==2.9
jwcrypto==0.7
md-toc==7.0.1
pathspec==0.7.0
pycparser==2.20
python-dateutil==2.8.1
regex==2020.2.20
requests==2.23.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
